// src/enums/monitoring.enum.ts
export enum MetaTypeEnum {
  AMAZON_PROFILE_ID = 'AMAZON_PROFILE_ID',
  ANARIX_ACCOUNT_ID = 'ANARIX_ACCOUNT_ID',
  AMAZON_ADVERTISER_ID = 'AMAZON_ADVERTISER_ID',
  AMAZON_SELLING_PARTNER_ID = 'AMAZON_SELLING_PARTNER_ID',
  WALMART_ADVERTISING_ID = 'WALMART_ADVERTISING_ID',
  WALMART_PARTNER_ID = 'WALMART_PARTNER_ID',
  ALL_ACCOUNT = 'ALL_ACCOUNT',
  AMAZON_DSP_ADVERTISER_ID = 'AMAZON_DSP_ADVERTISER_ID',
  AMAZON_AMC_ACCOUNT_ID = 'AMAZON_AMC_ACCOUNT_ID',
  WALMART_1P_SUPPLIER_ACCOUNT_ID = 'WALMART_1P_SUPPLIER_ACCOUNT_ID',
}

export enum MonitoringStatusEnum {
  RUNNING = 'RUNNING',
  PENDING = 'PENDING',
  FAILED = 'FAILED',
  COMPLETED = 'COMPLETED',
  STALE = 'STALE',
  STALE_RUNNING = 'STALE_RUNNING',
  UNEXPECTED = 'UNEXPECTED',
  CAPPED_OUT = 'CAPPED_OUT',
}

export enum ScheduleTypeEnum {
  SCHEDULED = 'SCHEDULED',
  AD_HOC = 'AD_HOC',
  TRIGGERED_BY_CODE = 'TRIGGERED_BY_CODE',
}

export enum MonitoringTableTitlesEnum {
  MONITORING_HOME = 'Monitoring',
  MONITORING_HISTORY = 'Monitoring_History',
  CRON_DEFINITIONS = 'Cron_Definitions',
  SQS_QUEUES = 'SQS_Queues',
}

export enum MonitoringSearchColumnsEnum {
  TASK_TYPE = 'taskType',
  TASK_ID = 'taskId',
}

export enum MonitoringMetaDataTypeEnum {
  INFO = 'INFO',
  SUCCESS = 'SUCCESS',
  ERROR = 'ERROR',
}

export enum SqsQueueNameEnum {
  SQS_DLQ_QUEUE_FIFO_URL = 'SQS_DLQ_QUEUE_FIFO_URL',
  SQS_JOB_QUEUE_FIFO_URL = 'SQS_JOB_QUEUE_FIFO_URL',
  SQS_DAYPARTING_QUEUE_FIFO_URL = 'SQS_DAYPARTING_QUEUE_FIFO_URL',
  SQS_SERP_QUEUE_FIFO_URL = 'SQS_SERP_QUEUE_FIFO_URL',
  SQS_TARGETING_ACTION_QUEUE_FIFO_URL = 'SQS_TARGETING_ACTION_QUEUE_FIFO_URL',
  SQS_AUTH_JOB_QUEUE_FIFO_URL = 'SQS_AUTH_JOB_QUEUE_FIFO_URL',
  SQS_MONITORING_QUEUE_FIFO_URL = 'SQS_MONITORING_QUEUE_FIFO_URL',
  SQS_AMAZON_ADS_QUEUE_FIFO_URL = 'SQS_AMAZON_ADS_QUEUE_FIFO_URL',
  SQS_WALMART_ADS_QUEUE_FIFO_URL = 'SQS_WALMART_ADS_QUEUE_FIFO_URL',
  SQS_BIDDER_QUEUE_FIFO_URL = 'SQS_BIDDER_QUEUE_FIFO_URL',
  SQS_AI_BIDDER_QUEUE_FIFO_URL = 'SQS_AI_BIDDER_QUEUE_FIFO_URL',
  SQS_AMC_QUEUE_FIFO_URL = 'SQS_AMC_QUEUE_FIFO_URL',
  SQS_DB_SYNC_QUEUE_FIFO_URL = 'SQS_DB_SYNC_QUEUE_FIFO_URL',
  SQS_AMAZON_SCRAPPER_QUEUE_FIFO_URL = 'SQS_AMAZON_SCRAPPER_QUEUE_FIFO_URL',
  SQS_RULES_QUEUE_FIFO_URL = 'SQS_RULES_QUEUE_FIFO_URL',
  SQS_AMS_QUEUE_FIFO_URL = 'SQS_AMS_QUEUE_FIFO_URL',
  SQS_MAIL_QUEUE_FIFO_URL = 'SQS_MAIL_QUEUE_FIFO_URL',
  /*
    SQS_AMAZON_SERP_STANDARD_URL is special queue with visibility timeout as 1.5 hour
    and this queue is only used for amazon scrapper, that too only for scrapping nothing else 
  */
  SQS_AMAZON_SERP_STANDARD_URL = 'SQS_AMAZON_SERP_STANDARD_URL',
  SQS_DLQ_STANDARD_URL = 'SQS_DLQ_STANDARD_URL',
  DUMMY_QUEUE = 'DUMMY_QUEUE',
}